pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sim-sandbox"
version = "0.1.0"
description = "对话大模型机器人仿真沙盒框架 — 基于 LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "typer[all]>=0.9",
    "jinja2>=3.1",
    "rich>=13.0",
    "numpy>=1.24",
    "langchain>=0.3",
    "langgraph>=0.2",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
sandbox = "sandbox.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sandbox"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
